Fix selection when two single-window frames
With two single-window frames, `next-window' was being called with
ALL-FRAMES as nil, resulting in the current window being returned
instead of the non-selected window. This commit sets the ALL-FRAMES
argument to 'visible' and modifies `aw-switch-to-window' to account for
the two-window case where POSITION may not be a window in the current
frame.

The only handler this commit modifies is `aw-switch-to-window' because
the behavior for `aw-delete-window' and `aw-swap-window' with multiple
frames doesn't seem to be well specified at this point.