Allow to change window action midway
* ace-window.el (aw--done): Update.
(aw-dispatch-function): New defvar. Set this to customize the midway
switch behavior.
(aw-action): New defvar. Midway switch will change this variable.
(aw-set-mode-line): New defun.
(aw-dispatch-default): New defun. Catch a char not in `aw-keys' and
perform an action depending on it.
(aw-select): Set `aw-action' to the ACTION arg, and call it in the
end. It can be changed during the selection.
(aw-split-window-vert): New command.
(aw-split-window-horz): New command.

Fixes #44