Fix missing leading characters
Fixes #75

The problem this commit fixes comes from the fact that sometimes emacs
scrolls the window when the overlay applied. This can happen for example
if the aw-leading-char-face's height is increased to make it bigger and
the cursor is near the bottom of the window. Then emacs might scroll the
window to make sure the cursor stays visible.
The idea here is to temporarily move the cursor to the leading char's
position for the overlay and then restore it when we're done with the
overlay.

Change-Id: I3921c3d4e26a6d5df4a431c8bbf4f2f3b12cb382