Make window ordering rely on frame position
I'm not quite sure what exactly the intention was of doing this off the id comparisons of the frames (they don't seem to be very consistent?).
I think that going off frame positions is much more intuitive. This only considers horizontal frame position, but it might also be worth considering their vertical position.