Fix window highlight in hollow cursor
When the cursor in non-selected windows becomes hollow, and the point is in the
same position as the window label, the label becomes invisible as it is rendered
with the background color of the cursor face.

Setting `cursor-in-non-selected-windows` to `nil` during selection, cursors in
non-selected windows are hidden completely. Hence the window label is always
visible.