Make aw-make-frame disable-able
There are times when I don't want a new frame made under any
circumstances. If that's the case, I don't want to accidentally
trigger aw-make-frame by hitting the wrong key at the dispatch.
Setting aw-make-frame-char to nil takes it off the key list, but this
requires a slight change to the dispatch check.