Implement alternative lead overlay creation strategy
An empty overlay is created when the buffer is empty (no need to
modify the buffer), the overlay is positioned before a tab (no
horizontal shift of the line when the character width of the lead face
differs from that of the underlying face) or it is positioned before a
new line. Additionally, only the path is fontified (no full width tab
or line when the lead face specifies a background) and a priority is
specified (no fighting with other overlays, see #98 and #178).

* ace-window.el
(aw--make-lead): New function.
(aw--done, aw--lead-overlay): Adapt accordingly.
(aw-empty-buffers-list, aw--overlay-str): Delete definitions.